+++ polls.js
@@ -7,12 +7,12 @@ module.exports.polls = function(parent) {
     obj.meshServer = parent.parent;
     obj.db = null;
 
-    obj.askQuestion = function(nodeid, question) {
-        obj.meshServer.sendToNode(nodeid, { action: 'polls_ask', question: question });
+    obj.askQuestion = function(nodeid, question, type, options) {
+        obj.meshServer.sendToNode(nodeid, { action: 'polls_ask', question: question, type: type, options: options });
     };
 
-    obj.answerQuestion = function(nodeid, question, answer) {
-        obj.db.run("INSERT INTO polls (nodeid, question, answer) VALUES (?, ?, ?)", [nodeid, question, answer]);
+    obj.answerQuestion = function(nodeid, question, answer, type, options) {
+        obj.db.run("INSERT INTO polls (nodeid, question, answer, type, options) VALUES (?, ?, ?, ?, ?)", [nodeid, question, answer, type, options]);
     };
 
     // Called when the plugin is loaded
@@ -20,12 +20,12 @@ module.exports.polls = function(parent) {
         obj.db = db;
 
         // Create the polls table if it doesn't exist
-        obj.db.run("CREATE TABLE IF NOT EXISTS polls (nodeid TEXT, question TEXT, answer TEXT)");
+        obj.db.run("CREATE TABLE IF NOT EXISTS polls (nodeid TEXT, question TEXT, answer TEXT, type TEXT, options TEXT)");
 
         obj.meshServer.on('agentCoreDirect', function(meshserver, module, data, nodeid) {
             if (module.name == 'polls') {
                 if (data.action == 'polls_answer') {
-                    obj.answerQuestion(nodeid, data.question, data.answer);
+                    obj.answerQuestion(nodeid, data.question, data.answer, data.type, data.options);
                 }
             }
         });
@@ -33,7 +33,7 @@ module.exports.polls = function(parent) {
         obj.meshServer.on('plugin', function(packet) {
             if (packet.plugin == 'polls') {
                 if (packet.command == 'ask') {
-                    obj.askQuestion(packet.data.nodeid, packet.data.question);
+                    obj.askQuestion(packet.data.nodeid, packet.data.question, packet.data.type, packet.data.options);
                 }
             }
         });