+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2 - 2025-07-24
+
+-   Added support for multiple choice and yes/no questions.
+-   Client UI now adapts to the question type.
+
 ## 0.0.1 - 2025-07-24
 
 -   Initial release